cmake: Use find_program for searching toolchain.
For Linux, it will try to pick up toolchain from $PATH,
and falls back to the toolchain in the utils folder
when it fails.
For macOS, it won't pass cmake's test when it fails back
to the utils folder. So the process will just fail if no
available toolchain is found from $PATH.